pyproject.toml
[tool.poetry]
name = "my_playbook_repo"
version = "0.0.1"
description = ""
authors = ["Xgrid <ali.khan@xgrid.co>"]

[tool.poetry.dependencies]
# if your playbook requires additional dependencies, add them here
#some-dependency = "^1.2.3"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
